Move onConnect/set_protocol to client/setProtocol
